Add files via upload
A GUI that I made to keep track of which row I am on while knitting and crocheting. It has a timer to keep track of how long I have been working on a project and an option to also keep track of which row of a repeat of x rows I am on while keeping track of the total number of rows. When the row number is changed it automatically prints the current row and time to a JTextArea to make it easy to copy elsewhere to track progress.